+++ Userland/Libraries/LibWeb/CMakeLists.txt
@@ -74,6 +74,7 @@ set(SOURCES
     CSS/StyleValues/ColorStyleValue.cpp
     CSS/StyleValues/ContentStyleValue.cpp
     CSS/StyleValues/FilterValueListStyleValue.cpp
+    CSS/StyleValues/FlexFlowStyleValue.cpp
     CSS/StyleValues/FlexStyleValue.cpp
     CSS/Supports.cpp
     CSS/SyntaxHighlighter/SyntaxHighlighter.cpp
